+++ src/systems/dnd5e.js
@@ -4,7 +4,7 @@ import CONSTANTS from "../constants/constants.js";
 
 export default {
 
-	"VERSION": "1.0.7",
+	"VERSION": "1.0.8",
 
 	// The actor class type is the type of actor that will be used for the default item pile actor that is created on first item drop.
 	"ACTOR_CLASS_TYPE": "character",
@@ -152,7 +152,7 @@ export default {
 	},
 
 	// Item similarities determines how item piles detect similarities and differences in the system
-	"ITEM_SIMILARITIES": ["name", "type"],
+	"ITEM_SIMILARITIES": ["name", "type", "system.container"],
 
 	// Currencies in item piles is a versatile system that can accept actor attributes (a number field on the actor's sheet) or items (actual items in their inventory)
 	// In the case of attributes, the path is relative to the "actor.system"